Se agregaron excepciones custom para Producto. Ahora dispara la excepcion productoNoEncontrado si no encuentra el ID.

+++ src/main/java/com/tpo/ecommerce/grupo6/service/ProductoService.java
@@ -1,12 +1,12 @@
 package com.tpo.ecommerce.grupo6.service;
 
+import com.tpo.ecommerce.grupo6.exception.ProductoNoEncontradoException;
 import com.tpo.ecommerce.grupo6.model.Producto;
 import com.tpo.ecommerce.grupo6.repository.ProductoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProductoService {
@@ -18,8 +18,9 @@ public class ProductoService {
         return productoRepository.findAll();
     }
 
-    public Optional<Producto> findById(Long id) {
-        return productoRepository.findById(id);
+    public Producto findById(Long id) {
+        return productoRepository.findById(id)
+                .orElseThrow(() -> new ProductoNoEncontradoException("Producto no encontrado con id: " + id));
     }
 
     public Producto save(Producto producto) {
@@ -27,6 +28,9 @@ public class ProductoService {
     }
 
     public void deleteById(Long id) {
+        if (!productoRepository.existsById(id)) {
+            throw new ProductoNoEncontradoException("No se puede eliminar. Producto no encontrado con id: " + id);
+        }
         productoRepository.deleteById(id);
     }
 }
